article page lang fixes
