+++ client/src/App.js
@@ -5,31 +5,33 @@ import Main from "./pages/Main.js";
 import About from "./About/About.js";
 import Header from "./Header/Header.js";
 import ArticlePage from "./ArticlePage/ArticlePage.js";
-import { myContext } from "./context/language.js";
+import ContextProvider,{ myContext } from "./context/language.js";
 
 
 function App() {
-  const { setTopArticles, URI } = useContext(myContext);
+  // const { setTopArticles, URI } = useContext(myContext);
 
-  const topArticles = async () => {
-    const articles = await axios.get(`${URI}`);
-    setTopArticles(articles.data);
-  };
+  // const topArticles = async () => {
+  //   const articles = await axios.get(`${URI}`);
+  //   setTopArticles(articles.data);
+  // };
 
-  useEffect(() => {
-    topArticles();
-  }, []);
+  // useEffect(() => {
+  //   topArticles();
+  // }, []);
 
   return (
-    <BrowserRouter forceRefresh={true}>
+    <ContextProvider>
+      <BrowserRouter forceRefresh={true}>
       <Header />
 
-      <Switch>
-        <Route path="/" exact component={Main} />
-        <Route path="/about" exact component={About} />
-        <Route path="/article/:id" exact component={ArticlePage} />
-      </Switch>
-    </BrowserRouter>
+        <Switch>
+          <Route path="/" exact component={Main} />
+          <Route path="/about" exact component={About} />
+          <Route path="/article/:id" exact component={ArticlePage} />
+        </Switch>
+        </BrowserRouter>
+      </ContextProvider>
   );
 }
