+++ client/src/index.js
@@ -1,13 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.js";
-import ContextProvider from "./context/language.js";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <ContextProvider>
       <App />
-    </ContextProvider>
   </React.StrictMode>
 );
